chore: move `rite_to_file` to `cargo_near_build::near::abi`
