+++ cargo-near-build/Cargo.toml
@@ -17,3 +17,4 @@ symbolic-debuginfo = "8.8"
 libloading = "0.8.3"
 serde_json = "1.0"
 near-abi = { version = "0.4.0", features = ["__chunked-entries"] }
+zstd = "0.13"
