+++ cargo-near-build/src/near/mod.rs
@@ -1,1 +1,2 @@
+// TODO: make mod non-pub
 pub mod abi;
